only use float32 in project_to_image

+++ rosys/vision/calibration.py
@@ -195,7 +195,7 @@ class Calibration:
         coordinates: np.ndarray = args[0]
         frame: Frame3d | None = kwargs.get('frame')
         world_extrinsics = self.extrinsics.relative_to(frame)
-        R = world_extrinsics.rotation.matrix
+        R = world_extrinsics.rotation.matrix.astype(np.float32)
         Rod = cv2.Rodrigues(R.T)[0]
         t = -R.T @ world_extrinsics.translation
         K = np.array(self.intrinsics.matrix)
